Committee on the Theory of Risk (COTOR)

The Casualty Actuarial Society's Committee on Theory of Risk (COTOR) issued several challenges as part of their work. These involved artificial constructed loss data. Actuaries were invited to construct models to estimate appropriate premiums for risk transfer.

Note that there is no COTOR1.

COTOR2

COTOR2 is a vector containing 250 observations of claim values.

COTOR round 2

COTOR3

COTOR3 is a data frame with 490 observations of 2 variables. The values were generated by the great Stuart Klugman.

Year: an integer indicating the year in which the claim occurred.

Amount: the amount of the claim.

COTOR round 3

COTOR4

COTOR4 is a data frame with 2500 rows and 2 columns.

Year: an integer indicating the year in which the claim occurred.

Amount: the amount of the claim.

COTOR round 4

COTOR5

COTOR5 is a data frame with 4849 rows and 5 columns

Risk: A single letter indicating the individual risk to which the claim belongs.

Year: an integer indicating the year in which the claim occurred.

MSA: simulated metropolitan statistical area.

Amount: the amount of the claim.

COTOR round 5

data(COTOR2)