New Jersey Manufacturing Company Workers Comp Reserving Data

This data represents ten complete years of Schedule P workers comp development for the New Jersey Manufacturing Company. The data was prepared by Glenn Meyers and Peng Shi and is available from the CAS https://www.casact.org/research/index.cfm?fa=loss_reserves_data. The columns have been given more user friendly names. The original column names may be restored by using the CasColNames function.

This data has been included because it is the triangle used by Greg Taylor and Grainne McGuire in their monograph "Stochastic Loss Reserving Using Generalized Linear Models". https://www.casact.org/pubs/monographs/index.cfm?fa=taylor-monograph03

This is a data frame with 100 observations of 13 variables.

GroupCode: NAIC company code

Company

AccidentYear

DevelopmentYear

Lag: Integer indicating the development lag in years.

CumulativeIncurred

CumulativePaid

IBNR: Bulk and IBNR reserves

DirectEP: Direct and assumed earned premium, gross of reinsurance.

CededEP: Earned premium ceded to reinsurers.

NetEP: Premium earned net of reinsurance.

Single: Either 1 or 0. 1 indicates a single entity, 0 indicates a group entity.

Reserve1997: Posted reserves as of year-end 1997.

data(NJM_WC)

Format

An object of class tbl_df (inherits from tbl, data.frame) with 100 rows and 13 columns.