Private Passenger Auto (PPA) Ratemaking Data

This is data taken from Appendix A of the "Basic Ratemaking" study note by Werner and Modlin.

Rate change

This is a data frame with 7 observations of 5 variables.

RateGroup: Letters from A-G. Note that rate group "A" does not have any rate changes.

EffectiveDate: The effective date of the rate change.

RateChange: A number indicating the rate change. Note that this is NOT done as a percent, i.e. .05 = 5

RateLevelIndex: This is one plus the rate change.

CumulativeRateLevel: This is the cumulative product of the RateLevelIndex. Note that this varies slightly from the figures shown in the study note. I believe this is due to rounding in the study note.

Accident year

This is a data frame with 5 observations of 4 variables.

AccidentYear: The accident year.

EarnedExposure: The exposure earned for the accident year.

EarnedPremium: The premium earned for the accieent year.

ReportedLoss: The reported loss and paid ALAE for the accident year.

Premium trend

Data frame with 25 observations of 3 values. The premium is shown at current level. This may be used to determine premium or exposure trend.

QuarterEnding: The ending quarter as a Date type.

WrittenPremium: Premium written during the quarter at current rate level.

WrittenExposure: Exposure written during the quarter.

Loss development

Data frame with 25 observations of 3 values. Note that this loss development data is stored in "long" rather than "wide" format.

AccidentYear: The accident year as an integer.

Lag: The lag in months as an integer.

ReportedLoss: The reported loss and paid ALAE for that accident year and lag.

Loss trend

Data frame with 20 observations of 7 values. The data may be used to estimate loss trend.

YearEndingQuarter: The end of a one-year period as a Date.

EarnedExposure: The exposure earned during the period.

ClosedClaimCount: The number of claims closed during the period.

PaidLosses: The amount of loss paid during the period.

Frequency: The closed claim count divided by the exposure.

Severity: The paid losses divided by the closed claim count.

PurePremium: The paid losses divided by the earned exposure.

ULAE

A data frame with 3 observations of 4 values.

CalendarYear: The calendar year as an integer.

PaidLoss_ALAE: The paid loss and ALAE for the calendar year.

PaidULAE: The ULAE paid in the calendar year.

ULAE_Ratio: The ratio of ULAE to the paid loss and ALAE.

data(PPA)